[DSY-292] Refactoring colors tab structure
