+++ sample/src/main/AndroidManifest.xml
@@ -63,7 +63,7 @@
             android:name=".components.StyleButtonActivity"
             android:theme="@style/Theme.Natura" />
         <activity
-            android:name=".components.colors.ColorsActivity"/>
+            android:name=".components.ColorsActivity"/>
     </application>
 
 </manifest>
