+++ sample/src/main/kotlin/com/natura/android/sample/MainActivity.kt
@@ -15,7 +15,7 @@ import com.natura.android.sample.components.StyleButtonActivity
 import com.natura.android.sample.components.SubmenuActivity
 import com.natura.android.sample.components.TextFieldActivity
 import com.natura.android.sample.components.ValueTextHighlightActivity
-import com.natura.android.sample.components.colors.ColorsActivity
+import com.natura.android.sample.components.ColorsActivity
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
