+++ sample/src/main/kotlin/com/natura/android/sample/components/ColorsActivity.kt
@@ -1,21 +1,18 @@
-package com.natura.android.sample.components.colors
+package com.natura.android.sample.components
 import android.content.Intent
-import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
-import androidx.viewpager.widget.ViewPager
 import com.google.android.material.tabs.TabLayout
 import com.natura.android.sample.R
-import com.natura.android.sample.setContentViewWithBrand
 import kotlinx.android.synthetic.main.activity_color.*
 
 class ColorsActivity : AppCompatActivity() {
 
-    var darkMode = false
-    var currentTab = 0
+    private var darkMode = false
+    private var currentTab = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -26,16 +23,13 @@ class ColorsActivity : AppCompatActivity() {
 
         setContentView(R.layout.activity_color)
         setSupportActionBar(colorToolbar)
-
-        brandsViewPager.adapter = ColorsTabAdapter(darkMode, supportFragmentManager)
-        brandsViewPager.currentItem = currentTab
-        colorTabs.setupWithViewPager(brandsViewPager)
-        if(darkMode) {
-            colorToolbar.background = ColorDrawable(resources.getColor(R.color.darkSurface))
-        }
-        setTabClickAction()
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        colorToolBar(darkMode)
 
+        val selectedTab = colorTabs.getTabAt(currentTab)
+        selectedTab?.select()
+
+        setTabClickAction()
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
@@ -46,24 +40,23 @@ class ColorsActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem?): Boolean {
         if(item?.itemId == R.id.darkThemeButton) {
             darkMode = !darkMode
-            currentTab = brandsViewPager.currentItem
-            val intent = Intent(baseContext, this.javaClass)
-            intent.putExtra("darkMode", darkMode)
-            intent.putExtra("currentTab", currentTab)
-            startActivity(intent)
+            restartScreenToApplyTheme(darkMode, currentTab)
         } else {
             onBackPressed()
         }
         return true
     }
 
+    private fun colorToolBar(darkMode: Boolean) {
+        if(darkMode) {
+            colorToolbar.background = ColorDrawable(resources.getColor(R.color.darkSurface))
+        }
+    }
+
     private fun setTabClickAction() {
         colorTabs.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
             override fun onTabSelected(tab: TabLayout.Tab?) {
-                val intent = Intent(baseContext, ColorsActivity::class.java)
-                intent.putExtra("darkMode", darkMode)
-                intent.putExtra("currentTab", tab?.position)
-                startActivity(intent)
+                restartScreenToApplyTheme(darkMode, tab?.position)
             }
 
             override fun onTabUnselected(tab: TabLayout.Tab?) {
@@ -75,6 +68,15 @@ class ColorsActivity : AppCompatActivity() {
         })
     }
 
+    private fun restartScreenToApplyTheme(darkMode: Boolean, currentTab: Int? = 0) {
+        val intent = Intent(baseContext, ColorsActivity::class.java)
+        intent.putExtra("darkMode", darkMode)
+        intent.putExtra("currentTab",currentTab)
+        intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION)
+        startActivity(intent)
+        finish()
+        overridePendingTransition(0, 0)
+    }
 
     private fun setGlobalTheme(darkMode: Boolean, currentTab: Int) {
         if(darkMode) {
@@ -90,7 +92,6 @@ class ColorsActivity : AppCompatActivity() {
                 else -> this.theme.applyStyle(R.style.Theme_BodyShop_NoActionBar, true)
             }
         }
-
     }
 
 }
