--- sample/src/main/kotlin/com/natura/android/sample/components/colors/AvonFragment.kt
@@ -1,15 +1,0 @@
-package com.natura.android.sample.components.colors
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.natura.android.sample.R
-
-class AvonFragment(private val darkMode:Boolean) : Fragment() {
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_avon, container, false)
-    }
-}
