+++ sample/src/main/res/layout/colors_variants_layout.xml
