--- sample/src/main/res/layout/fragment_thebodyshop.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_height="wrap_content"
-    android:layout_width="match_parent">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:background="?android:attr/colorBackground"
-        android:layout_marginBottom="10dp">
-        <include layout="@layout/primary_secondary_colors" />
-        <include layout="@layout/background_surface_colors" />
-        <include layout="@layout/content_light_colors"/>
-        <include layout="@layout/feedback_colors" />
-    </LinearLayout>
-</ScrollView>
